+++ resources/js/components/organisms/common/FeatureSlide/FeatureSlide.tsx
@@ -1,11 +1,11 @@
-import SlideButton from "@/components/organisms/home/FeatureSlide/SlideButton";
-import { FeatureSlideContainer } from "@/components/organisms/home/FeatureSlide/FeatureSlideStyles";
-import { PreviewType } from "@/components/organisms/home/FeatureSlide/types/Slide";
+import SlideButton from "@/components/organisms/common/FeatureSlide/SlideButton";
+import { FeatureSlideContainer } from "@/components/organisms/common/FeatureSlide/FeatureSlideStyles";
+import { PreviewType } from "@/components/organisms/common/FeatureSlide/types/Slide";
 import { useState } from "react";
 import {
     PREVIEW_SECTION,
     SLIDE_BUTTONS,
-} from "@/components/organisms/home/FeatureSlide/constants/SlideDetails";
+} from "@/components/organisms/common/FeatureSlide/constants/SlideDetails";
 
 interface Props {
     slideTime?: number;
