+++ resources/js/components/organisms/common/FeatureSlide/FeatureSlideStyles.tsx
