+++ resources/js/components/organisms/common/FeatureSlide/SlideButton/index.tsx
@@ -1,0 +1,3 @@
+import SlideButton from "@/components/organisms/common/FeatureSlide/SlideButton/SlideButton";
+
+export default SlideButton;
