+++ resources/js/components/organisms/common/FeatureSlide/__tests__/FeatureSlide.test.tsx
