+++ resources/js/components/organisms/common/FeatureSlide/index.tsx
@@ -1,0 +1,3 @@
+import FeatureSlide from "@/components/organisms/common/FeatureSlide/FeatureSlide";
+
+export default FeatureSlide;
