+++ resources/js/components/organisms/common/FeatureSlide/types/Slide.ts
