+++ resources/js/components/organisms/common/FeatureSlide/types/Style.ts
