+++ resources/js/components/organisms/common/WorldMap/WorldMap.tsx
