+++ resources/js/components/organisms/common/WorldMap/__tests__/WorldMap.cy.tsx
@@ -1,4 +1,4 @@
-import WorldMap from "@/components/organisms/home/WorldMap";
+import WorldMap from "@/components/organisms/common/WorldMap";
 import SVGUtils from "@/utils/SVGUtils";
 
 describe("WorldMap", () => {
