+++ resources/js/components/organisms/common/WorldMap/constants/MapPoints.ts
