+++ resources/js/components/organisms/common/WorldMap/types/Styles.ts
