+++ resources/js/components/organisms/home/Feature/FeatureStyles.tsx
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const FeatureContainer = styled.section`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    row-gap: 10rem;
+    width: 100%;
+    height: 100%;
+    min-height: 100vh;
+`;
+FeatureContainer.displayName = "FeatureContainer";
