+++ resources/js/components/organisms/home/Feature/index.tsx
@@ -1,0 +1,3 @@
+import Feature from "@/components/organisms/home/Feature/Feature";
+
+export default Feature;
