--- resources/js/components/organisms/home/FeatureSlide/SlideButton/index.tsx
@@ -1,3 +1,0 @@
-import SlideButton from "@/components/organisms/home/FeatureSlide/SlideButton/SlideButton";
-
-export default SlideButton;
