--- resources/js/components/organisms/home/FeatureSlide/index.tsx
@@ -1,3 +1,0 @@
-import FeatureSlide from "@/components/organisms/home/FeatureSlide/FeatureSlide";
-
-export default FeatureSlide;
