+++ resources/js/components/organisms/home/FinalPrompt/index.tsx
@@ -1,0 +1,3 @@
+import FinalPrompt from "@/components/organisms/home/FinalPrompt/FinalPrompt";
+
+export default FinalPrompt;
