+++ resources/js/components/organisms/home/Hero/index.tsx
@@ -1,0 +1,3 @@
+import Hero from "./Hero";
+
+export default Hero;
