+++ resources/js/pages/Home/Home.tsx
@@ -1,113 +1,18 @@
-import Button from "@/components/atoms/buttons/Button";
-import WorldMap from "@/components/organisms/home/WorldMap";
-import FeatureSlide from "@/components/organisms/home/FeatureSlide";
 import GuestLayout from "@/layouts/GuestLayout";
-import {
-    HomePageContainer,
-    Hero,
-    Feature,
-    FinalPrompt,
-} from "@/pages/Home/HomeStyles";
-import DownChevronIcon from "@/Icons/DownChevronIcon";
-import PublicImagesUtil from "@/utils/PublicImagesUtil";
-import { useRef } from "react";
+import { HomePageContainer } from "@/pages/Home/HomeStyles";
+import { useState } from "react";
+import Hero from "@/components/organisms/home/Hero";
+import Feature from "@/components/organisms/home/Feature";
+import FinalPrompt from "@/components/organisms/home/FinalPrompt";
 
 const Home = () => {
-    const featureSectionRef = useRef<HTMLDivElement | null>(null);
-
-    function scrollToFeaturesSection() {
-        featureSectionRef.current?.scrollIntoView?.({ behavior: "smooth" });
-    }
+    const [scrollToRef, setScrollToRef] = useState<HTMLElement | null>(null);
 
     return (
         <HomePageContainer>
-            <Hero>
-                <div className="hero-background">
-                    <WorldMap selectNum={100} selectionInterval={5} />
-                </div>
-
-                <section className="hero-intro">
-                    <header className="common-intro">
-                        <h1>
-                            Next generation web app for{" "}
-                            <span>
-                                <br />
-                                secure and private messaging
-                            </span>
-                        </h1>
-                        <h2>
-                            Easy, reliable, private messaging. Available
-                            worldwide
-                        </h2>
-                        <Button $variant="ghost" $borderSize="lg">
-                            Get Started
-                        </Button>
-                    </header>
-
-                    <footer>
-                        <div className="preview-opacity-gradient">
-                            <div className="preview-border-animation">
-                                <div className="preview-container">
-                                    <img
-                                        src={PublicImagesUtil.getPublicImage({
-                                            type: "assets",
-                                            name: "chat_page",
-                                        })}
-                                        alt="chat_page"
-                                    />
-                                </div>
-                            </div>
-                        </div>
-
-                        <button
-                            className="scroll-down"
-                            onClick={scrollToFeaturesSection}
-                        >
-                            <DownChevronIcon />
-                            scroll down
-                        </button>
-                    </footer>
-                </section>
-            </Hero>
-
-            <Feature ref={featureSectionRef}>
-                <header className="common-intro">
-                    <h1>
-                        Discover the technology that will
-                        <span>
-                            <br /> power your chat
-                        </span>
-                    </h1>
-                    <h4>
-                        We aim to optimise our tech to provide a secure and
-                        reliable experience
-                    </h4>
-                    <Button $variant="ghost" $borderSize="lg">
-                        Signup for free
-                    </Button>
-                </header>
-
-                <FeatureSlide />
-            </Feature>
-
-            <FinalPrompt>
-                <header className="common-intro">
-                    <h1>
-                        Get started now, everything you
-                        <span>
-                            <br /> need and more
-                        </span>
-                    </h1>
-                    <h4>
-                        Join one of the largest messaging platform out there
-                        with over 120 million registered users and 200 billion
-                        messages delivered per day for free!
-                    </h4>
-                    <Button $variant="ghost" $borderSize="lg">
-                        Get started for free
-                    </Button>
-                </header>
-            </FinalPrompt>
+            <Hero scrollTo={scrollToRef} />
+            <Feature setScrollToRef={setScrollToRef} />
+            <FinalPrompt />
         </HomePageContainer>
     );
 };
